Fix installation
